On branch main Your branch is ahead of 'origin/main' by 2 commits.   (use "git push" to publish your local commits)
Changes to be committed:
	modified:   nvm

Changes not staged for commit:
	modified:   awsdata/aws (untracked content)

Please enter the commit message for your changes. Lines starting
with '#' will be ignored, and an empty message aborts the commit.

Date:      Fri Aug 4 20:52:28 2023 +1000

On branch main
Your branch is ahead of 'origin/main' by 3 commits.
  (use "git push" to publish your local commits)

Changes to be committed:
	modified:   nvm

Changes not staged for commit:
	modified:   awsdata/aws (untracked content)