Please enter the commit message for your changes. Lines starting with '#' will be ignored, and an empty message aborts the commit.
On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
	modified:   nvm

Changes not staged for commit:
	modified:   awsdata/aws (untracked content)

Please enter the commit message for your changes. Lines starting
with '#' will be ignored, and an empty message aborts the commit.

Date:      Sat Aug 5 09:48:20 2023 +1000

On branch main
Your branch is ahead of 'origin/main' by 1 commit.
  (use "git push" to publish your local commits)

Changes to be committed:
	modified:   nvm

Changes not staged for commit:
	modified:   awsdata/aws (untracked content)